glue: catch BaseExecption in _safe_call
We want the actually catch here also KeyboardInterrupt,
which we use for handling SIGINT and SIGTERM.

I am not sure when we broke this, but must be a long time.
This was preventing SIGTERM and SIGINT to correctly handle
graceful termination by calling modules destroy functions.

https://python.readthedocs.io/en/v2.7.2/library/exceptions.html
https://docs.python.org/3/library/exceptions.html#exception-hierarchy

The class hierarchy is:

```
BaseException
...
 +-- SystemExit
 +-- KeyboardInterrupt
 +-- GeneratorExit
 +-- Exception
...
```

Relates to TFT-1729

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>