Send SIGTERM to children when interrupted
This should resolve the handling for cases where the children
do not handle our timeout signal or if the SIGTERM is ignored
by the children processes run.

If signal is received we explicitely send SIGTERM to children
processes. In case the children do not terminate in 30 seconds,
we send a SIGKILL to them.

Resolves TFT-1729

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>