Terminate all children processes
We did not send SIGTERM to all children processes in the process tree
hierarchy, only the first ones.

We have seen this in wild, and it should address issues with pipeline
cancellation in certain cases, like here:

https://artifacts.dev.testing-farm.io/949ab68b-80aa-4937-b797-c75368580486/pipeline.log

Note that:

```
[19:11:09] [W] Signal SIGUSR1 received
[19:11:09] [W] Sending SIGTERM to child process 'ansible-playboo' (PID 287)
[19:18:52] [W] [worker_0] [git-rawhide2slz05n6/tests/tests.yml] Results file work-tests.ymlzh_582o_/tests-yps1zwac/results.yml contains nothing under 'results' key
[19:18:52] [E] [test-schedule-runner] Exception raised: Pipeline timeout expired.
```

The cancellation of the pipeline took 7 minutes, as there was another
`ansible-playbook` process not killed in the process tree.
It can actually take a very long time, if the test gets stuck ...

I tried to build an worker image with a hotfix and the process is now correctly
killed and the pipeline is terminated shortly after:

https://artifacts.dev.testing-farm.io/73ce1759-595b-42eb-922d-2563161f7df8/pipeline.log

See:

```
[20:13:00] [W] Signal SIGUSR1 received
[20:13:00] [W] Sending SIGTERM to child process 'ansible-playboo' (PID 287)
[20:13:00] [W] Sending SIGTERM to child process 'ansible-playboo' (PID 414)
[20:13:00] [W] Sending SIGTERM to child process 'ssh' (PID 420)
[20:13:03] [W] [worker_0] [git-rawhide0hu_r6an/tests/tests.yml] Results file work-tests.ymly6j8f44w/tests-y7005ab8/results.yml contains nothing under 'results' key
[20:13:03] [E] [test-schedule-runner] Exception raised: Pipeline timeout expired.
```

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>