Add log topics support, add `eval-context` log topic
We want to add additional way to filter logging based on log topics.
We will use this to omit verbose `eval-context` logging by default,
it will be logged only if user requests logging of this log topic.

The logging is applicable to all log levels and thus all logging
functions.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>