glue: support ignoring errors during destroy
Because the pipeline blows up on the first traceback
of the `destroy` function, other vital destroy methods
of modules do not run and are causing outages of Testing Farm.

Introduce a new gluetool option `--ignore-destroy-failures`
to control if the `destroy` method failures should be ignored.
When ignored, all destroy functions will run and their failure
will be ignored by the pipelne. The failures will still be logged
and sent to Sentry.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>