utils.wait: display error message in GlueError
This is a recurring problem: our `gluetool.utils.wait` function
does not expose the error encountered when timeout happens.
This hides important information from the eyes of the users
and is taking our time to find the error in debug log.

To fix this, return the last encountered error from the `check`
function to provide more hints to the users.

Use `GlueError` exception message directly to print the error
message, so we can map it easily to a more user friendly error
if needed.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>